Add files via upload
Sara does not like strings which have both the uppercase and the lowercase characters. She thinks only those strings are good which have only lowercase characters(a-z) or only uppercase characters(A-Z).

Given a string s consisting of english alphabets. Find the minimum number of operations required to make it good.

In one operation you can choose any character of the string and change it into uppercase or lowercase.